ahmed

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -5,22 +5,22 @@
 /**
  * main -Entry point
  *
- * Description: A C progam that printswith printf function.
+ * Description: false or true validation
  *
  * Return: Always 0 (Success)
 */
 
 int main(void)
 {
-	float  n;
+	int  n;
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 	if (n > 0)
-		printf("%f is positvie\n", n);
+		printf("%i is positvie\n", n);
 	else if (n == 0)
-		printf("%f is zero\n", n);
+		printf("%i is zero\n", n);
 	else
-		printf("%f is negative\n", n);
+		printf("%i is negative\n", n);
 	return (0);
 }
